Adicionado a classe linha e iniciando as alteraçoes

+++ Exercicio13_5/src/udc/edu/exercicio13_5/FormaGeometrica.java
@@ -4,12 +4,12 @@ import udc.edu.exercicio13_5_1.Ponto2d;
 
 public abstract class FormaGeometrica {
 	public abstract Ponto2d centro();
-	public abstract float perimetro();
-	public abstract float area();
-	public abstract float largura();
-	public abstract float altura();
+	public abstract double perimetro();
+	public abstract double area();
+	public abstract double largura();
+	public abstract double altura();
 	
-	public float distancia(FormaGeometrica forma) {
+	public double distancia(FormaGeometrica forma) {
 		return centro().distancia(forma.centro());
 	}
